Convert feedback module to feedback service
